requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
boto3==1.26.76
coverage==7.3.4
junit-xml==1.9
lxml==4.9.4
numpy==1.26.0
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.0.2
psutil==5.9.7
py==1.11.0
pyspark==3.5.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-metadata==2.0.0
pytest-nunit==1.0.4
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-docx==1.1.0
requests==2.31.0
scikit-image==0.22.0
selenium==4.11.2
webdriver-manager==4.0.1
xlrd==2.0.1
psycopg2==2.9.9